extract plot_camera from camera module
